change directory structure and test structure
