+++ src/eval/num/complex.rs
@@ -1,5 +1,5 @@
 use super::number::{Arith, Zero};
-use crate::lisp::eval::expr::Eql;
+use crate::eval::expr::Eql;
 use std::fmt::Display;
 
 #[derive(Clone, Debug, PartialEq)]
