+++ src/eval/num/int.rs
@@ -1,5 +1,5 @@
 use super::number::{Arith, Compare, Zero};
-use crate::lisp::eval::expr::Eql;
+use crate::eval::expr::Eql;
 use std::fmt::Display;
 
 #[derive(Clone, Debug, PartialEq)]
