+++ src/eval/num/ratio.rs
@@ -1,6 +1,6 @@
 use super::math::gcd::gcd;
 use super::number::{Arith, Compare, Zero};
-use crate::lisp::eval::expr::Eql;
+use crate::eval::expr::Eql;
 use std::fmt::Display;
 
 #[derive(Clone, Debug, PartialEq)]
