+++ src/lib.rs
@@ -1,0 +1,2 @@
+#![allow(dead_code)]
+pub mod eval;
