+++ src/main.rs
@@ -1,7 +1,3 @@
-#![allow(dead_code)]
-
-mod lisp;
-
 fn main() {
     println!("Hello, world!");
 }
